made camera setup with image rig but copied everything to file on drive
I found the final MA file i used from last semester with all render layers set up. copied all cameras and settings for the tripo scene onto this and saved as a new file called Jake_TripSceneNotBearCircleNoReferencesOptimized.ma. this will be on google drive. pull this commit (the one you are reading right now) onto the school computer. and download the file from google drive and work on that from now on.